Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
alembic
pydantic
python-multipart 

# Thư viện bảo mật
passlib
bcrypt ==3.2.0

# TV cho JWT 
python-jose[cryptography] 